+++ crates/cli-support/src/wit/mod.rs
@@ -868,11 +868,13 @@ impl<'a> Context<'a> {
     fn string_enum(&mut self, string_enum: &decode::StringEnum<'_>) -> Result<(), Error> {
         let aux = AuxStringEnum {
             name: string_enum.name.to_string(),
+            comments: concatenate_comments(&string_enum.comments),
             variant_values: string_enum
                 .variant_values
                 .iter()
                 .map(|v| v.to_string())
                 .collect(),
+            generate_typescript: string_enum.generate_typescript,
         };
         let mut result = Ok(());
         self.aux